+++ lib/rack/handler/rage.rb
@@ -1,0 +1,2 @@
+require "rage"
+Rack::Handler.register("rage", "Rage::Handler") if defined?(Rack::Handler)