+++ lib/rackup/handler/rage.rb
@@ -1,0 +1,2 @@
+require "rage"
+Rackup::Handler.register("rage", "Rage::Handler") if defined?(Rackup::Handler)